relay_spans

Function derive_description_for_v2_span

pub fn derive_description_for_v2_span(
    attributes: &Attributes,
    name: &Annotated<String>,
) -> Option<String>

Derives a description for a V2 span, based on its name and attributes.

For now, this attempts to return the following values, in order:

the span’s name if its SENTRY__ORIGIN attribute is "manual"

a name constructed following the rules defined in sentry-conventions

None