relay_spans

Function derive_op_for_v2_span

pub fn derive_op_for_v2_span(attributes: &Annotated<Attributes>) -> String

Generates a sentry.op attribute for V2 span, if possible.

This uses attributes of the span to figure out an appropriate operation name, inferring what the SDK might have sent. Reliably infers an op for well-known OTel span kinds like database operations. Does not infer an op for frontend and mobile spans sent by Sentry SDKs that don’t have an OTel equivalent (e.g., resource loads).