relay_spans
Module otel_trace

Modules

span

Nested message and enum types in Span.

status

Nested message and enum types in Status.

Structs

ResourceSpans

A collection of ScopeSpans from a Resource.

ScopeSpans

A collection of Spans produced by an InstrumentationScope.

Span

A Span represents a single operation performed by a single component of the system.

Status

The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs.

TracesData

TracesData represents the traces data that can be stored in a persistent storage, OR can be embedded by other protocols that transfer OTLP traces data but do not implement the OTLP protocol.

Enums

SpanFlags

SpanFlags represents constants used to interpret the Span.flags field, which is protobuf ‘fixed32’ type and is to be used as bit-fields. Each non-zero value defined in this enum is a bit-mask. To extract the bit-field, for example, use an expression like: