Struct relay_spans::otel_trace::Span

pub struct Span {

    pub trace_id: Vec<u8>,
    pub span_id: Vec<u8>,
    pub trace_state: String,
    pub parent_span_id: Vec<u8>,
    pub flags: u32,
    pub name: String,
    pub kind: i32,
    pub start_time_unix_nano: u64,
    pub end_time_unix_nano: u64,
    pub attributes: Vec<KeyValue>,
    pub dropped_attributes_count: u32,
    pub events: Vec<Event>,
    pub dropped_events_count: u32,
    pub links: Vec<Link>,
    pub dropped_links_count: u32,
    pub status: Option<Status>,

}

A Span represents a single operation performed by a single component of the system.

The next available field id is 17.

Fields
trace_id: Vec<u8>

A unique identifier for a trace. All spans from the same trace share the same trace_id. The ID is a 16-byte array. An ID with all zeroes OR of length other than 16 bytes is considered invalid (empty string in OTLP/JSON is zero-length and thus is also invalid).

This field is required.

span_id: Vec<u8>

A unique identifier for a span within a trace, assigned when the span is created. The ID is an 8-byte array. An ID with all zeroes OR of length other than 8 bytes is considered invalid (empty string in OTLP/JSON is zero-length and thus is also invalid).

This field is required.

trace_state: String

trace_state conveys information about request position in multiple distributed tracing graphs. It is a trace_state in w3c-trace-context format: https://www.w3.org/TR/trace-context/#tracestate-header See also https://github.com/w3c/distributed-tracing for more details about this field.

parent_span_id: Vec<u8>

The span_id of this span’s parent span. If this is a root span, then this field must be empty. The ID is an 8-byte array.

flags: u32

Flags, a bit field. 8 least significant bits are the trace flags as defined in W3C Trace Context specification. Readers MUST not assume that 24 most significant bits will be zero. To read the 8-bit W3C trace flag, use flags & SPAN_FLAGS_TRACE_FLAGS_MASK.

When creating span messages, if the message is logically forwarded from another source with an equivalent flags fields (i.e., usually another OTLP span message), the field SHOULD be copied as-is. If creating from a source that does not have an equivalent flags field (such as a runtime representation of an OpenTelemetry span), the high 24 bits MUST be set to zero.

[Optional].

See https://www.w3.org/TR/trace-context-2/#trace-flags for the flag definitions.

name: String

A description of the span’s operation.

For example, the name can be a qualified method name or a file name and a line number where the operation is called. A best practice is to use the same display name at the same call point in an application. This makes it easier to correlate spans in different traces.

This field is semantically required to be set to non-empty string. Empty value is equivalent to an unknown span name.

This field is required.

kind: i32

Distinguishes between spans generated in a particular context. For example, two spans with the same name may be distinguished using CLIENT (caller) and SERVER (callee) to identify queueing latency associated with the span.

start_time_unix_nano: u64

start_time_unix_nano is the start time of the span. On the client side, this is the time kept by the local machine where the span execution starts. On the server side, this is the time when the server’s application handler starts running. Value is UNIX Epoch time in nanoseconds since 00:00:00 UTC on 1 January 1970.

This field is semantically required and it is expected that end_time >= start_time.

end_time_unix_nano: u64

end_time_unix_nano is the end time of the span. On the client side, this is the time kept by the local machine where the span execution ends. On the server side, this is the time when the server application handler stops running. Value is UNIX Epoch time in nanoseconds since 00:00:00 UTC on 1 January 1970.

This field is semantically required and it is expected that end_time >= start_time.

attributes: Vec<KeyValue>

attributes is a collection of key/value pairs. Note, global attributes like server name can be set using the resource API. Examples of attributes:

 "/http/user_agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_2) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/71.0.3578.98 Safari/537.36"
 "/http/server_latency": 300
 "example.com/myattribute": true
 "example.com/score": 10.239

The OpenTelemetry API specification further restricts the allowed value types: https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/common/README.md#attribute Attribute keys MUST be unique (it is not allowed to have more than one attribute with the same key).

dropped_attributes_count: u32

dropped_attributes_count is the number of attributes that were discarded. Attributes can be discarded because their keys are too long or because there are too many attributes. If this value is 0, then no attributes were dropped.

events: Vec<Event>

events is a collection of Event items.

dropped_events_count: u32

dropped_events_count is the number of dropped events. If the value is 0, then no events were dropped.

links: Vec<Link>

links is a collection of Links, which are references from this span to a span in the same or different trace.

dropped_links_count: u32

dropped_links_count is the number of dropped links after the maximum size was enforced. If this value is 0, then no links were dropped.

status: Option<Status>

An optional final status for this span. Semantically when Status isn’t set, it means span’s status code is unset, i.e. assume STATUS_CODE_UNSET (code = 0).

Implementations

impl Span

pub fn kind(&self) -> SpanKind

Returns the enum value of kind, or the default if the field is set to an invalid enum value.

pub fn set_kind(&mut self, value: SpanKind)

Sets kind to the provided enum value.