Crate relay_spans

Structs and functions needed to ingest OpenTelemetry spans.

Modules

otel_trace

Functions

name_for_span

Constructs a name attribute for a span, following the rules defined in sentry-conventions.

otel_to_sentry_span_v1

Transforms an OTEL span to a Sentry span.

otel_to_sentry_span_v2

Transform an OTEL span to a Sentry span V2.

span_v1_to_span_v2

Converts a legacy span to the new Span V2 schema.

span_v2_to_span_v1

Transforms a Sentry span V2 to a Sentry span V1.